requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
cffi==1.16.0
click==8.1.7
exceptiongroup==1.2.1
fastapi==0.110.3
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
starlette-graphene3==0.6.0
typing_extensions==4.11.0
uvicorn==0.29.0
